pyproject.toml
[project]
name = "llm-tester"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.0.0",
    "anthropic>=0.49.0",
    "openai>=1.70.0",
    "google-genai>=1.0.0",
    "matplotlib>=3.9.0",
]
